# perf: optimize homepage images to be max 300px